refactor(ui): correct module name typo
