+++ src/ui/update/mod.rs
@@ -1,6 +1,6 @@
 pub mod config;
-pub mod pomor;
+pub mod pomo;
 pub use config::ConfigCmd;
 pub use config::ConfigMsg;
-pub use pomor::PomodoroCmd;
-pub use pomor::PomodoroMsg;
+pub use pomo::PomodoroCmd;
+pub use pomo::PomodoroMsg;
